fix(script): decode CREATE tx arguments
