+++ cli/src/cmd/forge/script/broadcast.rs
@@ -1,10 +1,9 @@
-use super::{
-    sequence::{ScriptSequence, TransactionWithMetadata},
-    *,
-};
+use super::{sequence::ScriptSequence, *};
 use crate::{
     cmd::{
-        forge::script::{receipts::wait_for_receipts, verify::VerifyBundle},
+        forge::script::{
+            receipts::wait_for_receipts, transaction::TransactionWithMetadata, verify::VerifyBundle,
+        },
         has_batch_support, has_different_gas_calc,
     },
     init_progress,
